LPD-94550 Consolidate SEOStudioIntegrationRegistry into one class
