--- modules/dxp/apps/seo-studio/seo-studio-web/src/main/java/com/liferay/seo/studio/web/integration/SEOStudioIntegrationRegistry.java
@@ -1,19 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.seo.studio.web.integration;
-
-import java.util.Map;
-
-/**
- * @author Kiana Suetani
- */
-public interface SEOStudioIntegrationRegistry {
-
-	public SEOStudioIntegration getSEOStudioIntegration(String key);
-
-	public Map<String, SEOStudioIntegration> getSEOStudioIntegrations();
-
-}
